pyproject.toml
[project]
name = "enriched-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "dotenv>=0.9.9",
    "pydantic-ai>=0.3.3",
    "rich>=14.0.0",
]
